+++ src/main/scala/com/ebiznext/comet/job/index/bqload/BigQuerySparkJob.scala
@@ -18,7 +18,7 @@ import com.google.cloud.bigquery.{
 import com.google.cloud.hadoop.io.bigquery.BigQueryConfiguration
 import org.apache.hadoop.conf.Configuration
 import org.apache.spark.sql.functions.{col, date_format}
-import org.apache.spark.sql._
+import org.apache.spark.sql.{DataFrame, Row, SaveMode, SparkSession}
 import org.apache.spark.storage.StorageLevel
 
 import scala.collection.JavaConverters._
@@ -31,10 +31,7 @@ class BigQuerySparkJob(
     extends SparkJob
     with BigQueryJobBase {
 
-  /** Set extra spark conf here otherwise it will be too late once the spark session is created.
-    * @return
-    */
-  override def extraSparkConf(): Map[String, String] =
+  override def withExtraSparkConf(): Map[String, String] =
     if (cliConfig.writeDisposition == "WRITE_TRUNCATE")
       cliConfig.options ++ Map(
         "spark.datasource.bigquery.allowFieldAddition"   -> "false",
@@ -131,12 +128,6 @@ class BigQuerySparkJob(
   }
 
   def runSparkConnector(): Try[SparkJobResult] = {
-    def saveDfWithOptions(finalDF: DataFrameWriter[Row], saveMode: SaveMode) = {
-      logger.info(s"${cliConfig.options}")
-      cliConfig.options
-        .foldLeft(finalDF) { case (df, (k, v)) => df.option(k, v) }
-        .save()
-    }
     prepareConf()
     Try {
       val cacheStorageLevel =
@@ -199,9 +190,6 @@ class BigQuerySparkJob(
           cliConfig.partitionsToUpdate match {
             case None =>
             case Some(partitionsToUpdate) =>
-              logger.info(
-                s"Updating partitions $partitionsToUpdate overwriting partition ${partitionField} in The BQ Table $bqTable"
-              )
               partitionsToUpdate.foreach { partitionToUpdate =>
                 // if partitionToUpdate is not in the list of partitions to merge. It means that it need to be deleted
                 // this case happen when there is no more than a single element in the partition
@@ -215,7 +203,10 @@ class BigQuerySparkJob(
                     .option("datePartition", partitionToUpdate)
                     .option("table", bqTable)
                     .option("intermediateFormat", intermediateFormat)
-                  saveDfWithOptions(finalEmptyDF, SaveMode.Overwrite)
+
+                  cliConfig.options
+                    .foldLeft(finalEmptyDF) { case (df, (k, v)) => df.option(k, v) }
+                    .save()
                 }
               }
           }
@@ -231,16 +222,18 @@ class BigQuerySparkJob(
                 .option("datePartition", partitionStr)
                 .option("table", bqTable)
                 .option("intermediateFormat", intermediateFormat)
+            val finalDFWithOptions = cliConfig.options
+              .foldLeft(finalDF) { case (df, (k, v)) => df.option(k, v) }
             cliConfig.partitionsToUpdate match {
               case None =>
-                saveDfWithOptions(finalDF, SaveMode.Overwrite)
+                finalDFWithOptions.save()
               case Some(partitionsToUpdate) =>
                 // We only overwrite partitions containing updated or newly added elements
                 if (partitionsToUpdate.contains(partitionStr)) {
                   logger.info(
                     s"Optimization -> Writing partition : $partitionStr"
                   )
-                  saveDfWithOptions(finalDF, SaveMode.Overwrite)
+                  finalDFWithOptions.save()
                 } else {
                   logger.info(
                     s"Optimization -> Not writing partition : $partitionStr"
@@ -263,7 +256,8 @@ class BigQuerySparkJob(
             .format("com.google.cloud.spark.bigquery")
             .option("table", bqTable)
             .option("intermediateFormat", intermediateFormat)
-          saveDfWithOptions(finalDF, saveMode)
+
+          cliConfig.options.foldLeft(finalDF)((w, kv) => w.option(kv._1, kv._2)).save()
       }
 
       val stdTableDefinitionAfter =